requirements.txt
aiosmtplib==1.1.6
sentry-sdk==1.14.0
tenacity==8.2.1
jinja2==3.1.2
email-validator==1.3.1
asyncio==3.4.3
aioredis==2.0.1
APScheduler==3.9.1
transformers==4.32.0
openai==0.27.2
aiohttp==3.8.1
beautifulsoup4==4.10.0
backoff==1.11.1
ratelimit==2.2.1
scikit-learn==1.0.2
numpy==1.21.4
joblib==1.1.0
cryptography==3.4.8
boto3==1.20.24
python-dotenv==0.19.2
celery==5.2.3
redis==4.0.2
elasticsearch==7.16.1
pybreaker==0.8.0
docker==5.0.3
sendgrid==6.9.7
python-decouple==3.6
aiocache==0.11.1
prometheus_client==0.13.1
requests==2.26.0



